chat_backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
daphne
